+++ Proyecto-Core/ejbModule/com/core/service/negocio/remote/AdministradorEBR.java
@@ -1,6 +1,5 @@
 package com.core.service.negocio.remote;
 
-import java.math.BigDecimal;
 import java.util.Date;
 
 import javax.ejb.Remote;
@@ -10,7 +9,7 @@ public interface AdministradorEBR {
 	
 	public Boolean existeUsuario(String nick, String password);
 	public Long crearAdministrador(String nombre, String nick, String apellido,
-			String email, String password, Date fechaNac, String sexo, BigDecimal celular) throws Exception ;
+			String email, String password, Date fechaNac, String sexo, Integer celular) throws Exception ;
 	
 
 }
