+++ Proyecto-Core/ejbModule/com/core/service/negocio/remote/RescatistaEBR.java
@@ -12,7 +12,7 @@ import com.core.data.entites.Catastrofe;
 @Remote
 public interface RescatistaEBR {
 	
-	public Long crearRescatista(String nombre, String nick, String apellido, String email,String password,Date fechaNac,String sexo, BigDecimal celular) throws Exception;
+	public Long crearRescatista(String nombre, String nick, String apellido, String email,String password,Date fechaNac,String sexo, Integer celular) throws Exception;
 	
 	public void asignarRescatistaCatastrofe(Catastrofe catastrofe);
